requirements.txt
google-cloud-storage==1.40.0
google-cloud-secret-manager==2.5.0